Added lib/PPIrb.rb and began writing PPIrb.simpleBackup(backup)
lib/PPIrb.rb is intended to be the module that you interact with
when you want to be using ParanoidPackrat from Irb.  Also, when
you execute ./ParanoidPackrat.rb, it will be interacting with this
same library.

simpleBackup(backup) is exactly what it sounds like, a method that
performs a simple backup based on the 'backup' it is passed from
PPackratConfig[backupName].  More details on the process in the
comments in the source.